[type: refactor] refactor & optimize cache-plugin. (#3136)
* [type: refactor] refactor cache-plugin.

* code polish

* code polish
