+++ shenyu-plugin/shenyu-plugin-cache/pom.xml
@@ -30,12 +30,10 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>shenyu-plugin-cache-read</module>
-        <module>shenyu-plugin-cache-write</module>
         <module>shenyu-plugin-cache-spi</module>
         <module>shenyu-plugin-cache-memory</module>
         <module>shenyu-plugin-cache-redis</module>
-        <module>shenyu-plugin-cache-base</module>
+        <module>shenyu-plugin-cache-handler</module>
     </modules>
 
 </project>
