+++ shenyu-plugin/shenyu-plugin-global/src/test/java/org/apache/shenyu/plugin/global/GlobalPluginTest.java
@@ -82,7 +82,7 @@ public final class GlobalPluginTest {
 
     @Test
     public void testGetOrder() {
-        assertEquals(10, globalPlugin.getOrder());
+        assertEquals(5, globalPlugin.getOrder());
     }
 
     @Test
